refactor test cases with GrowFS on VhdXResize

+++ WS2012R2/lisa/remote-scripts/ica/STOR_VHDXResize_GrowFSAfterResize.sh
@@ -19,22 +19,26 @@
 # See the Apache Version 2.0 License for specific language governing
 # permissions and limitations under the License.
 #
-########################################################################
+######################################################################################
 # STOR_VHDXResize_GrowFSAfterResize.sh
 # Description:
-#     This script will verify if you can resize the filesystem on a resized  
-#     VHDx file
-#     Hyper-V setting pane. The test performs the following
-#     step
-#    1. Restores partition
-#    2. Checks the filesystem
-#    3. Perform filesystem resize
-#    4. Mounts the partition
-#    5. Performs read/write operations
-#    6. Unmounts partition
-#    7. Deletes partition  
-#
-########################################################################
+#  This script will verify if you can resize the filesystem on a resized VHDx file.
+#  The test performs the following steps:
+#  On first run:
+#    1. Create partition
+#    2. Format partition
+#    4. Mount the partition
+#    5. Read/Write mount point
+#    6. Unmount partition
+#  On second run (after vhdx is resized):
+#    4. Expand partition
+#    5. Check filesystem
+#    6. Resize filesystem
+#    4. Mount the partition
+#    5. Read/Write mount point
+#    6. Unmount partition
+#    7. Delete partition
+######################################################################################
 
 ICA_TESTRUNNING="TestRunning"      # The test is running
 ICA_TESTCOMPLETED="TestCompleted"  # The test completed successfully
@@ -60,7 +64,6 @@ if [ -e ~/summary.log ]; then
     rm -rf ~/summary.log
 fi
 
-LogMsg "Updating test case state to running"
 UpdateTestState $ICA_TESTRUNNING
 
 # Source the constants file
@@ -74,12 +77,11 @@ else
     exit 10
 fi
 
-#Prepare Read/Write script for execution
+# Prepare Read/Write script for execution
 dos2unix STOR_VHDXResize_ReadWrite.sh
 chmod +x STOR_VHDXResize_ReadWrite.sh
-#
+
 # Verify if guest sees the drive
-#
 if [ ! -e "/dev/sdb" ]; then
     msg="The Linux guest cannot detect the drive"
     LogMsg $msg
@@ -89,31 +91,29 @@ if [ ! -e "/dev/sdb" ]; then
 fi
 LogMsg "The Linux guest detected the drive"
 
-count=0
-for fs in "${fileSystems[@]}"; do
 
-    #
-    # Restore the partition
-    #
+if ! [ "$rerun" = "yes" ]; then
+    LogMsg "Info: Start testing filesystem: $fs"
+
+    # Create the partition
+    LogMsg "Info: Creating the partition on initial size of VHD."
     (echo n; echo p; echo 1; echo ; echo ;echo w) | fdisk /dev/sdb 2> /dev/null
     if [ $? -gt 0 ]; then
-        LogMsg "Failed to restore partition"
+        LogMsg "Failed to create partition."
         echo "Restore partition: Failed" >> ~/summary.log
         UpdateTestState $ICA_TESTFAILED
         exit 10
     fi
-    LogMsg "Partition restored"
+    LogMsg "Partition created."
 
+    partprobe
     sleep 5
 
-    # Format the partition
-    LogMsg "Info: Start testing filesystem: $fs"
+    # Verify if filesystem exist and then format the partition
     command -v mkfs.$fs
     if [ $? -ne 0 ]; then
         echo "Error: File-system tools for $fs not present. Skipping filesystem $fs.">> ~/summary.log
         LogMsg "Error: File-system tools for $fs not present. Skipping filesystem $fs."
-        count=`expr $count + 1`
-        continue
     else
         #Use -f option for xfs filesystem, but ignore parameter for other filesystems
         option=""
@@ -130,38 +130,87 @@ for fs in "${fileSystems[@]}"; do
         LogMsg "Info: Successfully formated partition with $fs"
     fi
 
-    #
+    # Mount partition
+    if [ ! -e "/mnt" ]; then
+        mkdir /mnt 2> ~/summary.log
+        if [ $? -ne 0 ]; then
+            LogMsg "Error: Failed to create mount point"
+            echo "Error: Creating mount point: Failed" >> ~/summary.log
+            UpdateTestState $ICA_TESTFAILED
+            exit 10
+        fi
+        LogMsg "Info: Mount point /dev/mnt created"
+    fi
+
+    mount /dev/sdb1 /mnt 2> ~/summary.log
+    if [ $? -ne 0 ]; then
+        LogMsg "Error: Failed to mount partition"
+        echo "Error: Mounting partition: Failed" >> ~/summary.log
+        UpdateTestState $ICA_TESTFAILED
+        exit 10
+    fi
+    LogMsg "Info: Partition mount successful"
+
+    mount | grep /dev/sdb1
+
+    # Read/Write mount point
+    ./STOR_VHDXResize_ReadWrite.sh
+
+    # Umount partition
+    umount /mnt
+    if [ $? -ne 0 ]; then
+        LogMsg "Error: Failed to unmount partition"
+        echo "Error: Unmounting partition: Failed" >> ~/summary.log
+        UpdateTestState $ICA_TESTFAILED
+        exit 10
+    fi
+    LogMsg "Info: Unmount partition successful"
+else
+    LogMsg "Info: Continue testing $fs."
+    LogMsg "Info: Expand partition to the new size"
+
+    # Expand partition to the new size of disk
+    (echo d; echo w) | fdisk /dev/sdb 2> /dev/null
+    partprobe
+    (echo n; echo p; echo 1; echo ; echo ;echo w) | fdisk /dev/sdb 2> /dev/null
+    if [ $? -gt 0 ]; then
+        LogMsg "Failed to expand partition"
+        echo "Expanding partition: Failed" >> ~/summary.log
+        UpdateTestState $ICA_TESTFAILED
+        exit 10
+    fi
+    LogMsg "Partition expanded"
+
+    partprobe
+    sleep 5
+
     # Checking filesystem
     # Because e2fsck and resize2fs only work for ext filesystems
     # we need to skip xfs for now
     if [ ! "$fs" = "xfs" ]; then
         e2fsck -y -v -f /dev/sdb1
         if [ $? -gt 0 ]; then
-            LogMsg "Failed to check filesystem"
-            echo "Checking filesystem: Failed" >> ~/summary.log
+            LogMsg "Failed to check filesystem $fs."
+            echo "Checking filesystem $fs: Failed" >> ~/summary.log
             UpdateTestState $ICA_TESTFAILED
             exit 10
         fi
-        LogMsg "Filesystem checked"
+        LogMsg "Filesystem $fs checked"
     fi
 
-    #
-    # Resizing the filesystem 
-    #
+    # Resizing the filesystem
     if [ ! "$fs" = "xfs" ]; then
         resize2fs /dev/sdb1
         if [ $? -gt 0 ]; then
-            LogMsg "Failed to resize filesystem"
-            echo "Resizing the filesystem: Failed" >> ~/summary.log
+            LogMsg "Failed to resize filesystem $fs."
+            echo "Resizing the filesystem $fs: Failed" >> ~/summary.log
             UpdateTestState $ICA_TESTFAILED
             exit 10
         fi
-    fi    
-    LogMsg "Filesystem resized"
+        LogMsg "Filesystem $fs resized"
+    fi
 
-    #
     # Mount partition
-    #
     if [ ! -e "/mnt" ]; then
         mkdir /mnt 2> ~/summary.log
         if [ $? -gt 0 ]; then
@@ -181,27 +230,25 @@ for fs in "${fileSystems[@]}"; do
         exit 10
     fi
     LogMsg "Partition mount successful"
-    
+    mount | grep /dev/sdb1
+
     # If the partition was successfully mounted we can use xfs_growsfs to
     # check the XFS filesystem also
     if [ "$fs" = "xfs" ]; then
-        xfs_growfs -d /dev/sdb1
+        xfs_growfs -d /mnt
         if [ $? -gt 0 ]; then
-            LogMsg "Failed to resize filesystem"
-            echo "Resizing the filesystem: Failed" >> ~/summary.log
+            LogMsg "Failed to resize filesystem $fs"
+            echo "Resizing the filesystem $fs: Failed" >> ~/summary.log
             UpdateTestState $ICA_TESTFAILED
             exit 10
         fi
+        LogMsg "Filesystem $fs resized"
     fi
 
-    #
     # Read/Write mount point
-    #
     ./STOR_VHDXResize_ReadWrite.sh
 
-    #
     # Restore ICA folder
-    #
     mkdir /mnt/ICA/ 2> ~/summary.log
     if [ $? -gt 0 ]; then
         LogMsg "Failed to restore directory /mnt/ICA/"
@@ -210,6 +257,7 @@ for fs in "${fileSystems[@]}"; do
         exit 10
     fi
 
+    # Umount partition
     umount /mnt
     if [ $? -gt 0 ]; then
         LogMsg "Failed to unmount partition"
@@ -218,16 +266,18 @@ for fs in "${fileSystems[@]}"; do
         exit 10
     fi
 
+    #Delete partition
     (echo d; echo w) | fdisk /dev/sdb 2> ~/summary.log
-    if [ $? -gt 0 ]; then
+    partprobe
+    lsblk | grep sdb1
+    if [ $? -eq 0 ]; then
         LogMsg "Failed to delete partition"
         echo "Deleting partition: Failed" >> ~/summary.log
         UpdateTestState $ICA_TESTFAILED
         exit 10
     fi
     LogMsg "Delete partition successful"
+    partprobe
+fi
 
-done
 UpdateTestState $ICA_TESTCOMPLETED
-
-exit 0;
